tooling/service-tags/requirements.txt
azure-identity>=1.15.0
azure-mgmt-network>=25.0.0
azure-mgmt-resource>=23.0.0
azure-mgmt-subscription>=3.1.1
azure-monitor-ingestion>=1.0.3